+++ app/models/itunes_image.rb
@@ -1,4 +1,6 @@
-class ITunesImage < PodcastImage
+class ITunesImage < BaseModel
+  include FeedImageFile
+
   validates :height, :width, numericality: {
     less_than_or_equal_to: 3000,
     greater_than_or_equal_to: 1400
@@ -7,8 +9,8 @@ class ITunesImage < PodcastImage
   validates :height, numericality: { equal_to: -> (image) { image.width } }, if: ->(i) { i.height }
 
   def replace_resources!
-    podcast.with_lock do
-      podcast.itunes_images.where("created_at < ? AND id != ?", created_at, id).destroy_all
+    feed.with_lock do
+      feed.itunes_images.where("created_at < ? AND id != ?", created_at, id).destroy_all
     end
   end
 end
