+++ test/factories/feed_image_factory.rb
@@ -1,12 +1,4 @@
 FactoryGirl.define do
-  factory :itunes_image do
-    original_url 'test/fixtures/valid_series_image.jpg'
-    width 1400
-    height 1400
-    size 95314
-    format 'jpeg'
-  end
-
   factory :feed_image do
     original_url 'test/fixtures/valid_feed_image.png'
     link 'http://www.maximumfun.org/jjgo'
